modified:   .gitignore 	modified:   Dockerfile 	modified:   README.md 	modified:   entrypoint.sh 	new file:   output.txt 	modified:   run.sh 	deleted:    run.txt 	new file:   screenshot.png 	modified:   server/index.html 	modified:   server/index.js 	new file:   server/model.txt 	deleted:    server/uploads/Badr Odhiambo_29a2930b-1a8d-4c59-af06-1b9643e9ffa3.wav.mp3 	deleted:    server/uploads/Claribel Dervla_c4d0edaa-eb8a-43dd-918b-b2917b3010c4.wav.mp3 	deleted:    server/uploads/Female-1_1b747d4c-9f41-4770-a3d1-82a1a71d278f.wav.mp3 	deleted:    server/uploads/Female-1_290db076-5381-4eb5-8c02-a00aac8310ec.wav.mp3 	deleted:    server/uploads/Female-1_534032a7-712d-481b-b15c-671d3af46c1d.wav.mp3 	deleted:    server/uploads/Female-1_83f140b4-aef6-433f-8add-0f3e007f4e5a.wav.mp3 	deleted:    server/uploads/Female-1_949f9495-ae2b-44f4-8936-08e80793c279.wav.mp3 	deleted:    server/uploads/Female-1_d0859622-e00e-492f-97fc-1daac47736c0.wav.mp3 	deleted:    server/uploads/Female-1_febc2e0e-7205-4410-ae51-99d13668ad69.wav.mp3 	deleted:    server/uploads/Female-2_d513f628-ffc0-48d5-9d82-884d87d41bad.wav.mp3 	deleted:    server/uploads/Henriette Usha_c74d7116-dbd5-4b2e-87f3-8061670912f5.wav.mp3 	deleted:    server/uploads/Henriette Usha_de874efe-3de1-424e-a74e-c1bb19d74776.wav.mp3 	deleted:    server/uploads/Male-1_739478dd-8665-4dab-970e-da97a8e321e0.wav.mp3 	deleted:    server/uploads/Male-1_c81294b3-901e-421c-b3e9-cc51ab96691f.wav.mp3 	deleted:    server/uploads/Male-4_5b8fd481-b288-46bd-a0bc-c3cf413cf0c2.wav.mp3 	deleted:    server/uploads/Male-7_db8c9796-9b94-4b56-bcf2-7075fa1341df.wav.mp3 	deleted:    server/uploads/Male-9_179510dc-ffed-4f12-a64c-04eeea104469.wav.mp3 	deleted:    server/uploads/Sofia Hellen_b92ec348-52fa-408c-b428-cf494e4c75f1.wav.mp3 	deleted:    server/uploads/Tammy Grit_5211a524-985f-42d7-afeb-23556f9bcce2.wav.mp3 	new file:   tempUploads/temporary-upload-directory

+++ server/index.js
@@ -1,58 +1,81 @@
+// Import necessary modules
 const express = require('express');
 const path = require('path');
 const fs = require('fs');
-const formidable = require('formidable');
+const { Formidable } = require('formidable');
 const { exec } = require('child_process');
 const keys = require('./keys.json');
 const { DateTime } = require('luxon');
 
+// Initialize express app
 const app = express();
 const port = 2901;
-const uploadDir = path.join(__dirname, 'uploads');
+const uploadDir = path.join('/shared', 'tempUploads');
+const model = fs.readFileSync(path.join('/shared', 'server', 'model.txt'), 'utf8').trim();
 
-// Middleware to parse JSON and urlencoded form data
+// Middleware to parse JSON and urlencoded form datasss
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
-// Route to handle form submission
-// Route for handling TTS requests
+/**
+ * Handle form submission
+ * @param {Object} req - The request object
+ * @param {Object} res - The response object
+ */
 app.post('/transcribe', async (req, res) => {
-    let { apiKey } = req.body;
+    const form = new Formidable({ multiples: false, uploadDir });
 
-    if (!authenticate(apiKey)) {
-        res.json({ status: 'error', message: 'Invalid API key.' });
-    }
-
-    const form = formidable({ multiples: false, uploadDir });
-
-    const user = keys.find(key => key.key === apiKey);
-    if (user) {
-        log(`${user.name} called.`);
-    }
-
-
+    // Parse form data
     form.parse(req, async (err, fields, files) => {
         if (err) {
             log(err);
             return res.send({ status: 'error', message: 'Error uploading audio file.' });
         }
 
+        const apiKey = fields.apiKey[0];
+
+        // Authenticate API key
+        if (!authenticate(apiKey)) {
+            return res.json({ status: 'error', message: 'Invalid API key: ' + apiKey + '' });
+        }
+
         // Get the uploaded audio file details
-        const audioFile = files.audio;
-        const filename = audioFile.name;
-        const fullFilePath = path.join(uploadDir, filename);
+        const audioFile = files.audio[0];
+        const fullFilePath = path.join('/shared', 'tempUploads', uuid() + audioFile.originalFilename);
 
-        const output = await transcribe(fullFilePath);
-        res.json({ status: 'success', output: output });
-        fs.unlinkSync(fullFilePath);
+        try {
+            // Check if source file exists
+            if (!fs.existsSync(audioFile.filepath)) {
+                log(`Source file does not exist: ${audioFile.filepath}`);
+                return res.json({ status: 'error', message: 'Source file does not exist.' });
+            }
+
+            // Check if destination directory exists, create if not
+            const destinationDir = path.dirname(fullFilePath);
+            if (!fs.existsSync(destinationDir)) {
+                log(`Destination directory does not exist, creating: ${destinationDir}`);
+                fs.mkdirSync(destinationDir, { recursive: true });
+            }
+
+            fs.renameSync(audioFile.filepath, fullFilePath);
+            const output = await transcribe(fullFilePath);
+            res.json({ status: 'success', message: fs.readFileSync(output, 'utf8') });
+            fs.unlinkSync(output);
+            fs.unlinkSync(fullFilePath);
+        } catch (error) {
+            log(error);
+            res.json({ status: 'error', message: 'An error occurred' });
+            fs.unlinkSync(audioFile.filepath);
+        }
     });
-
 });
 
+
+// Serve index.html
 app.get('/', (req, res) => {
-    // send index.html
     res.sendFile('/shared/server/index.html');
-})
+});
+
 
 // Start the server
 app.listen(port, () => {
@@ -60,33 +83,63 @@ app.listen(port, () => {
     log(`Server is running on http://localhost:${port}`);
 });
 
+/**
+ * Authenticate API key
+ * @param {string} apiKey - The API key to authenticate
+ * @returns {boolean} - True if API key is valid, false otherwise
+ */
 function authenticate(apiKey) {
     if (!apiKey) return false;
-    // get array of keys from keys.json
-    // check if apiKey is valid
     for (const key of keys) {
         if (key.key === apiKey) {
+            log(`${key.name} called API.`);
             return true;
         }
     }
     return false;
 }
 
-
-
+/**
+ * Log messages
+ * @param {string|Object} msg - The message to log
+ */
 function log(msg) {
     const today = DateTime.now().setZone("Africa/Lagos").toISODate();
-    fs.appendFileSync(`/shared/logs/${today}.log`, `${new Date().toLocaleString()}: ${JSON.stringify(msg)}\n`);
+    // if message is not string, convert it to string
+    if (typeof msg !== 'string') msg = JSON.stringify(msg);
+    fs.appendFileSync(`/shared/logs/${today}.log`, `${new Date().toLocaleString()}: ${msg}\n`);
 }
 
+
+
+/**
+ * Transcribe audio file
+ * @param {string} filePath - The path to the audio file
+ * @returns {Promise<string>} - The path to the transcript file
+ */
 function transcribe(filePath) {
+    const output = '/shared/tempUploads/' + uuid() + '.json';
+    const cmd = `insanely-fast-whisper --file-name ${filePath} --model-name ${model} --batch-size 24 --transcript-path ${output}`;
+
     return new Promise((resolve, reject) => {
-        exec(`insanely-fast-whisper --file-name "${filePath}" --flash True `, (error, stdout, stderr) => {
+        exec(cmd, (error, stdout, stderr) => {
             if (error) {
                 reject(error);
             } else {
-                resolve(stdout.trim()); // Trim any extra whitespace
+                resolve(output);
             }
         });
     });
 }
+
+/**
+ * Generate a UUID
+ * @returns {string} - The generated UUID
+ */
+function uuid() {
+    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
+        var r = Math.random() * 16 | 0, v = c == 'x' ? r : (r & 0x3 | 0x8);
+        return v.toString(16);
+    });
+}
+
